requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai>=1.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
# FastRTC和语音相关依赖
fastrtc[vad,tts]==0.0.25
google-generativeai>=0.8.0
numpy>=1.24.0
websockets>=12.0 